Add `api_key` to schema
I noticed when editing my prompty file in VS Code that it displayed a yellow
squiggly line under `api_key` because it wasn't in the schema. However, the
extension is using the field, because executing my prompty fails without it and
succeeds with it.

I added it to the schema so it's documented and the squiggly line goes away.