Add env var to disable image parsing
Added a `PROMPTY_CHAT_PARSER_DISABLE_IMAGE_PARSING=true` environment variable to
disable image parsing in the chat parser.

In my use case, I am using Prompty to send a prompt to the LLM that contains Markdown that
should be converted to a different format. I do not need Prompty to parse the
images and try to send them to the LLM. In fact, this is inconvenient because I don't
want to have to copy the images just so Prompty can read them.

An environment variable seems kind of janky perhaps, but it's the best solution
I could come up with that doesn't require a lot of changes, because there's a
fairly deep call stack so in order to pass a parameter, it would require
modifying a lot of code. I'm open to other ideas though.